LAB_1_3/application/requirements.txt
Flask
psycopg2_binary